requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
aiosmtplib==2.0.2
aiofiles==23.2.1
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
bcrypt==4.2.0
blinker==1.8.2
boto3==1.35.76
botocore==1.35.76
build==1.2.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
deprecation==2.1.0
distro==1.9.0
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.0
fastapi-mail==1.4.1
frozenlist==1.4.1
gotrue==2.9.1
greenlet==3.1.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.6
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
Mako==1.3.5
MarkupSafe==2.1.5
multidict==6.1.0
openai==1.51.0
packaging==24.1
passlib==1.7.4
pip-tools==7.4.1
pluggy==1.5.0
postgrest==0.17.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
pyproject_hooks==1.2.0
pytest==8.3.3
pytest-asyncio==0.24.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.12
realtime==2.0.5
requests==2.32.3
resend==2.4.0
rsa==4.9
s3transfer==0.10.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
sqlmodel==0.0.22
starlette==0.38.6
storage3==0.8.0
StrEnum==0.4.15
stripe==11.1.0
supabase==2.8.1
supafunc==0.6.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.31.0
websockets==13.1
yarl==1.13.1
